+++ src/info_provider.rs
@@ -1,6 +1,5 @@
 use nautilus_extension::{FileInfo, InfoProvider};
-use std::path::Path;
-use std::process::Command;
+use tmsu_commands;
 use url;
 
 pub struct TmsuInfoProvider {
@@ -14,24 +13,8 @@ impl InfoProvider for TmsuInfoProvider {
 
     fn update_file_info(&self, file_info: &mut FileInfo) {
         let path = get_path(file_info);
-
-        let output = Command::new("tmsu")
-                             .arg("tags")
-                             .arg(&path)
-                             .current_dir(Path::new(&path).parent().unwrap())
-                             .output()
-                             .expect("could not query tmsu tags");
-
-        let output_string = String::from_utf8(output.stdout.into_iter().collect()).unwrap();
-        let output_str = output_string.trim();
-        let find_result = output_str.find(": ");
-
         let attr_name = "tmsu_tags".to_string();
-        let attr_value =
-            match find_result {
-                Some(idx) => String::from(&output_str[idx+2..]),
-                None => String::new(),
-            };
+        let attr_value = tmsu_commands::tags(&path);
 
         file_info.attributes.insert(attr_name, attr_value);
     }
